requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
chardet==5.2.0
click==8.1.7
cssselect==1.2.0
dicttoxml==1.7.16
dj-database-url==2.2.0
Django==4.2.16
django-allauth==0.57.2
django-background-tasks==1.2.8
frozenlist==1.5.0
future==1.0.0
geocoder==1.38.1
geoip2==4.8.0
googlemaps==4.10.0
gunicorn==23.0.0
IP2Location==8.10.4
lazy-object-proxy==1.10.0
lxml==5.3.0
maxminddb==2.6.2
multidict==6.1.0
oauthlib==3.2.2
outcome==1.3.0.post0
pip-review==1.3.0
propcache==0.2.0
psycopg2==2.9.10
PyJWT==2.9.0
pyparsing==3.2.0
pyquery==2.0.1
PySocks==1.7.1
python3-openid==3.2.0
ratelim==0.1.6
requests-oauthlib==2.0.0
selenium==4.25.0
sortedcontainers==2.4.0
sqlparse==0.5.1
tqdm==4.66.5
trio==0.27.0
trio-websocket==0.11.1
typed_ast==1.5.5
typing==3.7.4.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.16.0
